specify dtype of feature extractor np array to be float

+++ plastering/data_feature_extractor.py
@@ -187,7 +187,7 @@ def haar_transform(x):
 class data_feature_extractor():
 
     def __init__(self, X):
-        self.X = np.asarray(X)
+        self.X = np.asarray(X, dtype=float)
         self.functions = [
         'getF_1994_Li',
         'getF_2012_Calbimonte',
